Update docs to manage expectations
There is no reason to worry about coercing parameters to Pinboard. It
will start off as an extremely thin (but complete) wrapper around the
API. This means that the choices made for argument types in the API will
be reflected in Thumbtack for a bit. Parameter type coercion will be
added later.